Adds Localization API.
To localize a plugin use,

```php
<?php
function localize() {
  $this->__('Foo string to localize');
}
```

The `local` localization helpers are `_`, `__` and `t`. These are
aliases to the `translate` method. The `translate` method takes a string
and localizes it using the WordPress `__` api method.

The text domain of a plugin is assumed to be the same as it's slug by
default.